moved code from usertest to posttest
